pyproject.toml
[project]
name = "mcp-annual-report"
version = "0.1.0"
description = "FastMCP server for annual report analysis with Bearer Token authentication"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.5.2",
    "python-dotenv>=1.1.0",
    "starlette>=0.47.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.12",
]

# Ruff 配置
[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "UP007", # Use `X | Y` for type annotations
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["mcp_annual_report"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_annual_report"]

[tool.hatch.build.targets.wheel.force-include]
"." = "."
